Add custom exceptions

+++ SolidCleanArchitectureCourse.Application/Features/LeaveType/Commands/DeleteLeaveType/DeleteLeaveTypeCommandHandler.cs
@@ -1,5 +1,6 @@
 ﻿using MediatR;
 using SolidCleanArchitectureCourse.Application.Contracts.Persistence;
+using SolidCleanArchitectureCourse.Application.Exceptions;
 
 namespace SolidCleanArchitectureCourse.Application.Features.LeaveType.Commands.DeleteLeaveType;
 
@@ -15,6 +16,10 @@ public class DeleteLeaveTypeCommandHandler : IRequestHandler<DeleteLeaveTypeComm
     public async Task<Unit> Handle(DeleteLeaveTypeCommand request, CancellationToken cancellationToken)
     {
         var leaveTypeToDelete = await _leaveTypeRepository.GetByIdAsync(request.Id);
+
+        if (leaveTypeToDelete is null)
+            throw new NotFoundException(nameof(Domain.LeaveType), request.Id);
+
         await _leaveTypeRepository.DeleteAsync(leaveTypeToDelete);
         return Unit.Value;
     }
